pyproject.toml
[tool.poetry]
name = "fever"
version = "0.1.0"
description = "Fever code challenge"
authors = ["Kata <kata@appears.gz>"]
readme = "README.md"
packages = [
    { include = "app", from = "." }
]


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
httpx = "^0.27.0"
pre-commit = "^3.7.0"
flake8 = "^7.0.0"
black = "^24.4.0"
mypy = "^1.9.0"
aiohttp = "^3.9.4"
fastapi-scheduler = "^0.0.15"
testcontainers = "^4.4.0"
pydantic-settings = "^2.2.1"
pydantic-xml = "^2.9.2"
tenacity = "^8.2.3"
pybreaker = "^1.2.0"
arq = "^0.26.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.23.6"
aioresponses = "^0.7.6"
coverage = "^7.4.4"
requests = "*"
ipykernel = "^6.29.4"
respx = "^0.21.1"
types-redis = "^4.6.0.20240425"
types-pyyaml = "^6.0.12.20240311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
markers = ["integration: marks tests as integration (deselect with '-m \"not integration\"')"]


[tool.black]
line-length = 120
target-version = ["py311"]
exclude = '''
/(
    \.git
    | \.venv
)/
'''

[tool.isort]
profile = "black"
src_paths = ["app"]

[tool.mypy]
check_untyped_defs = true
plugins = ["pydantic.mypy"]
strict = true
exclude = ["^.*test_.*\\.py$", "^.*conftest\\.py$", "docker/.*"]


[virtualenvs]
in-project = true
